refactor: move dns package to resolver package
