+++ pkg/gateway/engine_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/nite-coder/bifrost/pkg/config"
-	"github.com/nite-coder/bifrost/pkg/dns"
+	"github.com/nite-coder/bifrost/pkg/resolver"
 	"github.com/nite-coder/bifrost/pkg/variable"
 	"github.com/stretchr/testify/assert"
 )
@@ -23,7 +23,7 @@ func TestMiddlewarePipeline(t *testing.T) {
 		ServiceID: "testService",
 	}
 
-	dnsResolver, err := dns.NewResolver(dns.Options{})
+	dnsResolver, err := resolver.NewResolver(resolver.Options{})
 	assert.NoError(t, err)
 
 	bifrost := &Bifrost{
