+++ pkg/gateway/service_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/nite-coder/bifrost/pkg/config"
-	"github.com/nite-coder/bifrost/pkg/dns"
+	"github.com/nite-coder/bifrost/pkg/resolver"
 
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/cloudwego/hertz/pkg/app/server"
@@ -50,7 +50,7 @@ func TestServices(t *testing.T) {
 		_ = h.Shutdown(ctx)
 	}()
 
-	dnsResolver, err := dns.NewResolver(dns.Options{})
+	dnsResolver, err := resolver.NewResolver(resolver.Options{})
 	assert.NoError(t, err)
 
 	bifrost := &Bifrost{
@@ -133,7 +133,7 @@ func TestDynamicService(t *testing.T) {
 		_ = h.Shutdown(ctx)
 	}()
 
-	dnsResolver, err := dns.NewResolver(dns.Options{})
+	dnsResolver, err := resolver.NewResolver(resolver.Options{})
 	assert.NoError(t, err)
 
 	bifrost := &Bifrost{
@@ -180,7 +180,7 @@ func TestDynamicService(t *testing.T) {
 }
 
 func TestDynamicServiceMiddleware(t *testing.T) {
-	dnsResolver, err := dns.NewResolver(dns.Options{})
+	dnsResolver, err := resolver.NewResolver(resolver.Options{})
 	assert.NoError(t, err)
 
 	bifrost := &Bifrost{
