+++ pkg/resolver/resolver_test.go
@@ -1,4 +1,4 @@
-package dns
+package resolver
 
 import (
 	"context"
